[#749] Fix MEMORY/GTMASSERT2/JNLCNTRL/REPLJNLCLOSED errors from huge transactions in journaled database files
* See https://gitlab.com/YottaDB/DB/YDB/-/issues/749#description for more background on the issue.

* The main fix is to convert all places where `jgbl.cumul_jnl_rec_len` (which maintains the current
  total journal record size requirements of all replicated journal records in the current transaction)
  is incremented to additionally check if the total becomes greater than or equal to 1GiB. If so,
  we issue a new `REPLTRANSJNL1GB` error. This prevents the various issues that were seen before.
  A new macro `INCREMENT_JGBL_CUMUL_JNL_REC_LEN` takes care of this.

  This limit was originally set at `2GiB` in a former version of this commit. But we noticed various
  test failures. This is because replication code (e.g. `gtmsource_get_jnlrecs()`) has various length
  related fields hardcoded as type `int` instead of `unsigned int` and various functions returns a
  `-1` value of length to signify an error code path and does `0 > len` in various places. All of
  these will not work right if we support a `2GiB` max length as in some places we allocate twice
  this max length which will make the buffer length greater than `2GiB` and cause the `int` value to
  be treated as a negative value instead of a positive value. To avoid these issues, the limit was
  cut to half of `2GiB` and hence the `1GiB` limit.

* In `sr_port/tp_tend.c` and `sr_port/t_end.c`, the types of the `total_jnl_rec_size` and `tot_jrec_size`
  variables were changed from `uint4` to `gtm_uint64_t` (i.e. 4-byte to a 8-byte unsigned value) and
  various calculations are now typecast as `(uint4)` to avoid values greater than `2GiB` but less than
  `4GiB` to be treated as a negative value. Similar changes were also done in `sr_port/tp.h` to the macro
  `TOTAL_TPJNL_REC_SIZE`. This ensures the correct check happens for the `JNLTRANS2BIG` error.

  With the previous checks, it was possible we missed out on issuing a `JNLTRANS2BIG` error due to
  the value overflowing past `4GiB` and circling back to a small positive 4-byte value. We would
  then proceed to commit this transaction and `rsrv_freeaddr` would end up getting bumped to a value
  greater than `4GiB` which would cause it to circle back to a small positive 4-byte value and then
  result in a situation where `jb->freeaddr` is less than `jb->dskaddr` which is an out-of-design
  situation and results in the `JNLCNTRL` and/or `REPLJNLCLOSED` errors.

* `sr_port/tp_tend.c` : `si->tot_jrec_size` field was replaced with a local variable `tot_jrec_size`
  and removed from the `sgm_info` structure in `sr_port/tp.h` as this was the only use of it.

* Changed the type of the global variable `gtmsource_cmpmsgbufsiz` from `int` to `unsigned int` to allow
  greater than `2Gib` (but less than `4Gib`) sizes since the compressed buffer can be at most twice
  the regular buffer size (which can be greater than `1GiB` but less than `2GiB`).

  Because of this, an assert in `sr_unix/gtm_zlib.h` was changed to just check for non-zero values (since
  the check for being positive was okay previously for a `int` type but is no longer appropriate for an
  `unsigned int` type).

+++ sr_unix/gtmsource_process_ops.c
@@ -99,7 +99,7 @@ GBLREF	FILE			*gtmsource_log_fp;
 GBLREF	gd_addr			*gd_header;
 GBLREF	gtmsource_state_t	gtmsource_state;
 GBLREF	int4			strm_index;
-GBLREF	int			gtmsource_cmpmsgbufsiz;
+GBLREF	unsigned int		gtmsource_cmpmsgbufsiz;
 GBLREF	int			gtmsource_filter;
 GBLREF	int			gtmsource_log_fd;
 GBLREF	int			gtmsource_msgbufsiz;
@@ -366,10 +366,10 @@ int gtmsource_alloc_msgbuff(int maxbuffsize, boolean_t discard_oldbuff)
 			 * Allocate extra space just in case compression actually expands the data (needed only in rare cases).
 			 */
 			oldmsgp = gtmsource_cmpmsgp;
-			gtmsource_cmpmsgp = (repl_msg_ptr_t)malloc(maxbuffsize * MAX_CMP_EXPAND_FACTOR);
+			gtmsource_cmpmsgp = (repl_msg_ptr_t)malloc((unsigned)maxbuffsize * MAX_CMP_EXPAND_FACTOR);
 			if (NULL != oldmsgp)
 				free(oldmsgp);
-			gtmsource_cmpmsgbufsiz = (maxbuffsize * MAX_CMP_EXPAND_FACTOR);
+			gtmsource_cmpmsgbufsiz = ((unsigned)maxbuffsize * MAX_CMP_EXPAND_FACTOR);
 		}
 		gtmsource_alloc_filter_buff(gtmsource_msgbufsiz);
 	}
